update test data manger
